backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
python-dotenv
pydantic
pymongo
faker
mcp[cli]
langchain-groq
pyodbc
langchain-mcp-adapters
psycopg2-binary
